Bugfix re: POSIX find not having -printf, bugfix re: race condition causing disk full, bugfix re: POSIX date command
The race condition occurred because a pipeline had an output redirection to a
file name that could match a file glob at the beginning of the pipeline.  cat
very happily opened its own output file, and intent on emptying it AND filling
it... filled the disk.  Fixed by to the glob pattern to prevent the output
file from matching the input glob.